Cleaned up app.css, Bugfix in DbFileController
- Started to clean up app.css. Not done.
- Delete Properties in DbFileController.Delete.

+++ Caps.Web.UI/Controllers/DbFileController.cs
@@ -21,7 +21,11 @@ namespace Caps.Web.UI.Controllers
 
         public HttpResponseMessage Delete(int id)
         {
-            var file = db.Files.Include("Versions.Content").Include("Versions.Thumbnails").FirstOrDefault(f => f.Id == id);
+            var file = db.Files
+                .Include("Versions.Content")
+                .Include("Versions.Thumbnails")
+                .Include("Versions.Properties")
+                .FirstOrDefault(f => f.Id == id);
             if (file == null)
                 return Request.CreateResponse(HttpStatusCode.NotFound);
 
@@ -29,8 +33,8 @@ namespace Caps.Web.UI.Controllers
             {
                 Array.ForEach(file.Versions.ToArray(), v =>
                 {
-                    Array.ForEach(v.Thumbnails.ToArray(), t => db.Thumbnails.Remove(t));
-                    Array.ForEach(v.Properties.ToArray(), p => db.FileProperties.Remove(p));
+                    if (v.Thumbnails != null) Array.ForEach(v.Thumbnails.ToArray(), t => db.Thumbnails.Remove(t));
+                    if (v.Properties != null) Array.ForEach(v.Properties.ToArray(), p => db.FileProperties.Remove(p));
                     db.FileContents.Remove(v.Content);
                     db.FileVersions.Remove(v);
                 });
